requirements.txt
pandas
faiss-cpu
sentence_transformers==2.2.2
google-generativeai
keras==2.11.0
tf-keras
flask